refactor(api): move database connection to its own provider
