+++ apps/api/src/persistence/database/database.provider.spec.ts
@@ -1,34 +1,47 @@
 import { ConfigService } from '@nestjs/config';
 import { Test } from '@nestjs/testing';
+import { ArangoConnectionProvider } from './arango-connection.provider';
+import { ArangoDatabase } from './arango-database';
 import { DatabaseProvider } from './database.provider';
 
 describe('AppController', () => {
   let databaseProvider: DatabaseProvider;
   let configService: ConfigService;
 
-  let db;
+  let arangoInstance;
 
   beforeAll(async () => {
     const moduleRef = await Test.createTestingModule({
-      providers: [ConfigService],
+      providers: [ConfigService, ArangoConnectionProvider],
     }).compile();
 
     configService = moduleRef.get<ConfigService>(ConfigService);
     if (!configService) throw new Error('Config service not injected.');
-    databaseProvider = new DatabaseProvider(configService);
 
-    db = databaseProvider.getConnection();
+    const arangoConnectionProvider = moduleRef.get<ArangoConnectionProvider>(
+      ArangoConnectionProvider
+    );
+
+    if (!arangoConnectionProvider)
+      throw new Error('Connection provider not injected');
+
+    const databaseProvider = new DatabaseProvider(
+      configService,
+      arangoConnectionProvider
+    );
+
+    arangoInstance = await databaseProvider.getDBInstance();
   });
 
-  describe('getConnection', () => {
-    let result;
+  describe('get database instance', () => {
+    describe('the returned instance', () => {
+      it('should not be null or undefined', () => {
+        expect(arangoInstance).toBeTruthy();
+      });
 
-    beforeAll(async () => {
-      result = await db.route('_api').get('version');
-    });
-
-    it('querying the db version should return a result', () => {
-      expect(result).toBeTruthy();
+      it('should be an instance of ArangoDatabase', () => {
+        expect(arangoInstance).toBeInstanceOf(ArangoDatabase);
+      });
     });
   });
 });
