+++ apps/api/src/persistence/repositories/repository.provider.spec.ts
@@ -1,5 +1,6 @@
 import { ConfigService } from '@nestjs/config';
 import { Test } from '@nestjs/testing';
+import { ArangoConnectionProvider } from '../database/arango-connection.provider';
 import { DatabaseProvider } from '../database/database.provider';
 import { RepositoryProvider } from './repository.provider';
 
@@ -10,7 +11,7 @@ describe('Repository Provider', () => {
 
   beforeAll(async () => {
     const moduleRef = await Test.createTestingModule({
-      providers: [DatabaseProvider, ConfigService],
+      providers: [DatabaseProvider, ConfigService, ArangoConnectionProvider],
     }).compile();
 
     databaseProvier = moduleRef.get<DatabaseProvider>(DatabaseProvider);
